기능 추가, 개선 및 오류 수정  - useQuery 리팩토링  - ROLE 권한 관리 기능 추가  - 관리자 로그인 시 최초 진입 페이지 모니터링으로 변경  - 가리기 기능 개선  - 멀티화면 오류 수정  - 캘린더 오류 수정  - 알림 메뉴 모두 읽음 버튼 위치 변경  - 실현손익 최근 기준으로 정렬하도록 수정  - snackBar를 alert 형태로 전체 변경  - 각종 오류 수정

+++ src/api/calendar/EconomicCalendarApi.ts
@@ -18,23 +18,23 @@ export const fetchCalendarEvents = async (req: CalendarEventsReq) => {
     return res.data;
 };
 
-export const fetchManualCalendarEvents = async (req: CalendarEventsReq) => {
-    const res = await api.get('/calendar/events/manual', {params: req});
+export const fetchManualCalendarEvents = async (req: CalendarEventsReq, config?: AxiosRequestConfig) => {
+    const res = await api.get('/admin/calendar/events', {...config, params: req});
     return res.data;
 };
 
 export const createCalendarEvent = async (req: ManualCalendarEventReq) => {
-    const res = await api.post('/calendar/events', req);
+    const res = await api.post('/admin/calendar/events', req);
     return res.data;
 };
 
 export const updateCalendarEvent = async (id: number, req: ManualCalendarEventReq) => {
-    const res = await api.put(`/calendar/events/${id}`, req);
+    const res = await api.put(`/admin/calendar/events/${id}`, req);
     return res.data;
 };
 
 export const deleteCalendarEvent = async (id: number) => {
-    const res = await api.delete(`/calendar/events/${id}`);
+    const res = await api.delete(`/admin/calendar/events/${id}`);
     return res.data;
 };
